feat(MyGames): Make page mobile friendly
