+++ packages/client/src/modules/play/MyGames/index.ts
@@ -1,0 +1,1 @@
+export * from "./MyGames";
